+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.9.6 [vNext]
+- Generated imports that would end in `/index.js` (or ts/d.ts) are no longer simplified to the directory name in ESM
+  mode.
+- Generated imports that would begin with `@types/` are now simplified to a bare import for the module name (ie
+  `@types/express/foo/index.js` would become `express`).
+
 # v0.9.5
 - Fix cases where references to generated imports were for the default export instead of appropriate specific exports
   when targeting ES modules.